city_rag/requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
openai
python-dotenv
mysql-connector-python
